--- src/components/Testimonials/data.js
@@ -1,89 +1,0 @@
-export const data=[[
-    {
-        message:
-          `Never thought programming would be that easy and interesting. Just watched the binary search video of Kunal and literally understood every single thing. It's such a boon to learn from you. Thnak you so much for providing this help`,
-        img: '',
-        name: 'Vasihnavi Bholane',
-        position: 'Student'
-      },
-      {
-        message:
-          `I have learned Java in my second year and none of my professors have taught what Kunal has taught. I have learned the memory management in stack and heap memory`,
-        img: '',
-        name: 'mrr.ayman',
-        position: 'Student'
-      },
-      {
-        message:
-          `I am following Community Classroom's bootcamp and found it really useful. I am getting all the concepts taught so far crystal clear as well as able to solve leetcode questions given in the assignments. Before this I was finding leetcode very difficult.`,
-        img: '',
-        name: 'Anurag Pathak',
-        position: 'Student'
-      }],
-      [
-        {
-            message:
-              `Never thought programming would be that easy and interesting. Just watched the binary search video of Kunal and literally understood every single thing. It's such a boon to learn from you. Thnak you so much for providing this help`,
-            img: '',
-            name: 'Vasihnavi Bholane',
-            position: 'Student'
-          },
-          {
-            message:
-              `I have learned Java in my second year and none of my professors have taught what Kunal has taught. I have learned the memory management in stack and heap memory`,
-            img: '',
-            name: 'mrr.ayman',
-            position: 'Student'
-          },
-          {
-            message:
-              `I am following Community Classroom's bootcamp and found it really useful. I am getting all the concepts taught so far crystal clear as well as able to solve leetcode questions given in the assignments. Before this I was finding leetcode very difficult.`,
-            img: '',
-            name: 'Anurag Pathak',
-            position: 'Student'
-          }],
-          [
-            {
-                message:
-                  `Never thought programming would be that easy and interesting. Just watched the binary search video of Kunal and literally understood every single thing. It's such a boon to learn from you. Thnak you so much for providing this help`,
-                img: '',
-                name: 'Vasihnavi Bholane',
-                position: 'Student'
-              },
-              {
-                message:
-                  `I have learned Java in my second year and none of my professors have taught what Kunal has taught. I have learned the memory management in stack and heap memory`,
-                img: '',
-                name: 'mrr.ayman',
-                position: 'Student'
-              },
-              {
-                message:
-                  `I am following Community Classroom's bootcamp and found it really useful. I am getting all the concepts taught so far crystal clear as well as able to solve leetcode questions given in the assignments. Before this I was finding leetcode very difficult.`,
-                img: '',
-                name: 'Anurag Pathak',
-                position: 'Student'
-              }],
-              [
-                {
-                    message:
-                      `Never thought programming would be that easy and interesting. Just watched the binary search video of Kunal and literally understood every single thing. It's such a boon to learn from you. Thnak you so much for providing this help`,
-                    img: '',
-                    name: 'Vasihnavi Bholane',
-                    position: 'Student'
-                  },
-                  {
-                    message:
-                      `I have learned Java in my second year and none of my professors have taught what Kunal has taught. I have learned the memory management in stack and heap memory`,
-                    img: '',
-                    name: 'mrr.ayman',
-                    position: 'Student'
-                  },
-                  {
-                    message:
-                      `I am following Community Classroom's bootcamp and found it really useful. I am getting all the concepts taught so far crystal clear as well as able to solve leetcode questions given in the assignments. Before this I was finding leetcode very difficult.`,
-                    img: '',
-                    name: 'Anurag Pathak',
-                    position: 'Student'
-                  }],
-]
