+++ src/data/HomeViewData/homeviewdata.js
@@ -81,33 +81,105 @@ export const feature = [
     img: '/static'
   }
 ];
-export const students = [
+
+export const data=[[
   {
-    message:
-      'Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum ',
-    img: '',
-    name: 'Dummy',
-    position: 'Student'
-  },
-  {
-    message:
-      'Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum ',
-    img: '',
-    name: 'Dummy',
-    position: 'Student'
-  },
-  {
-    message:
-      'Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum ',
-    img: '',
-    name: 'Dummy',
-    position: 'Student'
-  },
-  {
-    message:
-      'Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem Ispum Lorem ispum ',
-    img: '',
-    name: 'Dummy',
-    position: 'Student'
-  }
-];
+      id: '1',
+      message:
+        `Never thought programming would be that easy and interesting. Just watched the binary search video of Kunal and literally understood every single thing. It's such a boon to learn from you. Thnak you so much for providing this help`,
+      img: '',
+      name: 'Vasihnavi Bholane',
+      position: 'Student'
+    },
+    {
+      id: '2',
+      message:
+        `I have learned Java in my second year and none of my professors have taught what Kunal has taught. I have learned the memory management in stack and heap memory`,
+      img: '',
+      name: 'mrr.ayman',
+      position: 'Student'
+    },
+    {
+      id: '3',
+      message:
+        `I am following Community Classroom's bootcamp and found it really useful. I am getting all the concepts taught so far crystal clear as well as able to solve leetcode questions given in the assignments. Before this I was finding leetcode very difficult.`,
+      img: '',
+      name: 'Anurag Pathak',
+      position: 'Student'
+    }],
+    [
+      {
+        id: '1',
+        message:
+            `Never thought programming would be that easy and interesting. Just watched the binary search video of Kunal and literally understood every single thing. It's such a boon to learn from you. Thnak you so much for providing this help`,
+          img: '',
+          name: 'Vasihnavi Bholane',
+          position: 'Student'
+        },
+        {
+          id: '2',
+          message:
+            `I have learned Java in my second year and none of my professors have taught what Kunal has taught. I have learned the memory management in stack and heap memory`,
+          img: '',
+          name: 'mrr.ayman',
+          position: 'Student'
+        },
+        {
+          id: '3',
+          message:
+            `I am following Community Classroom's bootcamp and found it really useful. I am getting all the concepts taught so far crystal clear as well as able to solve leetcode questions given in the assignments. Before this I was finding leetcode very difficult.`,
+          img: '',
+          name: 'Anurag Pathak',
+          position: 'Student'
+        }],
+        [
+          {
+            id: '1',
+            message:
+                `Never thought programming would be that easy and interesting. Just watched the binary search video of Kunal and literally understood every single thing. It's such a boon to learn from you. Thnak you so much for providing this help`,
+              img: '',
+              name: 'Vasihnavi Bholane',
+              position: 'Student'
+            },
+            {
+              id: '2',
+              message:
+                `I have learned Java in my second year and none of my professors have taught what Kunal has taught. I have learned the memory management in stack and heap memory`,
+              img: '',
+              name: 'mrr.ayman',
+              position: 'Student'
+            },
+            {
+              id: '3',
+              message:
+                `I am following Community Classroom's bootcamp and found it really useful. I am getting all the concepts taught so far crystal clear as well as able to solve leetcode questions given in the assignments. Before this I was finding leetcode very difficult.`,
+              img: '',
+              name: 'Anurag Pathak',
+              position: 'Student'
+            }],
+            [
+              {
+                id: '1',
+                message:
+                    `Never thought programming would be that easy and interesting. Just watched the binary search video of Kunal and literally understood every single thing. It's such a boon to learn from you. Thnak you so much for providing this help`,
+                  img: '',
+                  name: 'Vasihnavi Bholane',
+                  position: 'Student'
+                },
+                {
+                  id: '2',
+                  message:
+                    `I have learned Java in my second year and none of my professors have taught what Kunal has taught. I have learned the memory management in stack and heap memory`,
+                  img: '',
+                  name: 'mrr.ayman',
+                  position: 'Student'
+                },
+                {
+                  id: '3',
+                  message:
+                    `I am following Community Classroom's bootcamp and found it really useful. I am getting all the concepts taught so far crystal clear as well as able to solve leetcode questions given in the assignments. Before this I was finding leetcode very difficult.`,
+                  img: '',
+                  name: 'Anurag Pathak',
+                  position: 'Student'
+                }],
+]
